rust: add bitmap API.
Provides an abstraction for C bitmap API and bitops operations.
We follow the C Bitmap API closely in naming and semantics, with
a few differences that take advantage of Rust language facilities
and idioms:

  * We leverage Rust type system guarantees as follows:

    * Ownership transfer of a Bitmap is restricted so that it cannot
    be passed between threads (does not implement `Send`).

    * all (non-atomic) mutating operations require a &mut reference which
    amounts to exclusive access.

    * It is permissible to pass shared references &Bitmap between
    threads, and we expose atomic operations through interior mutability.
    Since these atomic operations do not provide any ordering guarantees,
    we mark these as `unsafe`. Anyone who calls the atomic methods needs
    to document that the lack of ordering is safe.

  * The Rust API offers `{set,clear}_bit` vs `{set,clear}_bit_atomic`,
    which is different from the C naming convention (set_bit vs __set_bit).

  * we include enough operations for the API to be useful, but not all
    operations are exposed yet in order to avoid dead code. This commit
    includes enough to enable a Rust implementation of an Android Binder
    data structure that was introduced in commit 15d9da3f818c ("binder:
    use bitmap for faster descriptor lookup"), which can be found in
    drivers/android/dbitmap.h. We need this in order to upstream the Rust
    port of Android Binder driver.

  * We follow the C API closely and fine-grained approach to safety:

    * Low-level bit-ops methods get a safe API with bounds checks.

    * methods correspond to find_* C methods tolerate out-of-bounds
    arguments. Since these are already safe we the same behavior, and
    return results using `Option` types to represent "not found".

  * the Rust API is optimized to represent the bitmap inline if it would
    take the space of a pointer. This saves allocations which is
    relevant in the Binder use case.

The underlying C bitmap is *not* exposed, and must never be exposed
(except in tests). Exposing the representation would lose all static
guarantees, and moreover would prevent the optimized representation of
short bitmaps.

An alternative route of vendoring an existing Rust bitmap package was
considered but suboptimal overall. Reusing the C implementation is
preferable for a basic data structure like bitmaps. It enables Rust
code to be a lot more similar and predictable with respect to C code
that uses the same data structures and enables the use of code that
has been tried-and-tested in the kernel, with the same performance
characteristics whenever possible.

We use the `usize` type for sizes and indices into the bitmap,
because Rust generally always uses that type for indices and lengths
and it will be more convenient if the API accepts that type. This means
that we need to perform some casts to/from u32 and usize, since the C
headers use unsigned int instead of size_t/unsigned long for these
numbers in some places.

Adds new MAINTAINERS section BITMAP API [RUST].

Suggested-by: Alice Ryhl <aliceryhl@google.com>
Suggested-by: Yury Norov <yury.norov@gmail.com>
Signed-off-by: Burak Emir <bqe@google.com>