rust: add bitmap API.
Provides an abstraction for C bitmap API and bitops operations.
We follow the C Bitmap API closely in naming and semantics, with
a few differences that take advantage of Rust language facilities
and idioms:

  * not all operations are exposed yet, to avoid dead code. This commit
    includes enough to implement an Android Binder data structure that
    was introduced in commit 15d9da3f818c ("binder: use bitmap for
    faster descriptor lookup"), namely drivers/android/dbitmap.h. This
    is part of upstreaming the Android Binder driver.

  * API uses Option types where appropriate

  * bound checks are used to treat out-of-bounds access as bug
    (hardening). The C operations treat out-of-bounds parameters
    as a default case e.g. "not found" which is safe (avoids UB) but
    may hide bugs.

  * better naming convention to distinguish non-atomic from atomic
    operations: {set,clear}_bit vs {set,clear}_bit_atomic.
    The Rust type system ensures that all non-atomic operations
    require a &mut reference which amounts to exclusive access.
    Using the atomic variants only makes sense when multiple threads
    have a shared reference &bitmap which amounts to the interior
    mutability pattern.

  * optimized to represent the bitmap inline if it would take the space
    of a pointer. This saves allocations which is relevant in the
    Binder use case.

The underlying C bitmap is *not* exposed. This would lose all static
guarantees.

An alternative route of vendoring an existing Rust bitmap package was
considered but suboptimal overall. Reusing the C implementation is
preferable for a basic data structure like bitmaps. It enables Rust
code to be a lot more similar and predictable with respect to C code
that uses the same data structures and enables the use of code that
has been tried-and-tested in the kernel.

We use the `usize` type for sizes and indices into the bitmap,
because Rust generally always uses that type for indices and lengths
and it will be more convenient if the API accepts that type. This means
that we need to perform some casts to/from u32 and usize, since the C
headers use unsigned int instead of size_t/unsigned long for these
numbers in some places.

Adds new MAINTAINERS section BITMAP API [RUST].

Suggested-by: Alice Ryhl <aliceryhl@google.com>
Suggested-by: Yury Norov <yury.norov@gmail.com>
Signed-off-by: Burak Emir <bqe@google.com>